+++ ambra_sdk/__init__.py
@@ -1,8 +1,8 @@
-__version__ = '3.20.3.0-5'
+__version__ = '3.20.4.0-1'
 
 # from footer in api.html
-API_VERSION = 'LBL0022 v37.0 2020-04-15'
+API_VERSION = 'LBL0022 v38.0 2020-05-27'
 # equal for API_VERSION based on dictionary.html
-MODELS_VERSION = 'LBL0022 v37.0 2020-04-15'
+MODELS_VERSION = 'LBL0022 v38.0 2020-05-27'
 # from storage_api.html
-STORAGE_VERSION = 'LBL0038 v8.0 2019-07-17'
+STORAGE_VERSION = 'LBL0038 v9.0 2020-06-03'