+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.20.4.0-1] - 2020-06-03
+### Added
+- Storage api support: LBL0038 v9.0 2020-06-03
+- Service public api support: LBL0022 v38.0 2020-05-27
+- Service models: LBL0022 v38.0 2020-05-27
+- get_tags and tag_by_name methods in Storage.Study.json Box response
+- get_tags and tag_by_name methods in Storage.Study.image_json Box response
+- Add InconsistencyConfict(409) error Storage.Study.schema
+
+### Changed
+- Box response type in storage methods now is optional (default return Box)
+
 ## [3.20.3.0-5] - 2020-05-06
 ### Added
 - Doctests in sphinx documentation
@@ -39,6 +51,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [3.20.3.0-3] - 2020-04-13
 ### Added
 - Storage api support: LBL0038 v8.0 2019-07-17
-- Service public api support: 
+- Service public api support: LBL0022 v37.0 2020-04-15
 - Service models: LBL0022 v37.0 2020-04-15
 - Docker support: LBL0022 v37.0 2020-04-15